Added name rendering to player

+++ src/com/dazkins/triad/gfx/Font.java
@@ -34,7 +34,7 @@ public class Font {
 		}
 	}
 	
-	public static void drawString(String s, int xp ,int yp) {
+	public static void drawString(String s, float xp ,float yp) {
 		String lower = s.toLowerCase();
 		
 		for (int x = 0; x < lower.length(); x++) {
@@ -48,7 +48,7 @@ public class Font {
 		}
 	}
 	
-	public static void drawStringWithShadow(String s, int xp ,int yp) {
+	public static void drawStringWithShadow(String s, float xp ,float yp) {
 		String lower = s.toLowerCase();
 		
 		for (int x = 0; x < lower.length(); x++) {
